feat(tech-inights): adds getFactSchemas method to client
